requirements.txt
Flask==2.2.2
flask-session==0.4.0
mysql-connector-python==8.0.32
requests==2.28.2
streamlit==1.22.0
werkzeug==2.2.2
